Rename SPNG_FMT_PNG to SPNG_FMT_RAW
This is more accurate since we aren't going to upscale the data
that is less than 8bits to 8bits. This makes handling 1,2,4bit data
a bit harder but is a better route since the end user doesn't have
to repack the bits if that's what they wanted in the first place.
(i.e. read PNG and immediately save as something else).